azure_developer_associate/irisapp/irisapi/pyproject.toml
[tool.poetry]
name = "irisapi"
version = "0.1.0"
description = ""
authors = ["Manuel Gil <manuelgilsitio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.116.1"}
scikit-learn = "^1.7.0"
azure-cosmos = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
